requirements.txt
# Volatility 3 Memory Forensics Platform - Python Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0  # PostgreSQL async driver
alembic==1.12.1  # Database migrations

# Task Queue
celery[redis]==5.3.4
redis==5.0.1
kombu==5.3.4

# Authentication & Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==41.0.7

# Validation & Serialization
pydantic==2.5.2
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Monitoring & Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing FastAPI

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# Production
gunicorn==21.2.0
